`tmpnet`: Break config.go up into coherent parts (#2462)
